+++ imports/ui/validators/Delegations.jsx
@@ -14,7 +14,7 @@ export default class ValidatorDelegations extends Component{
         this.state = {
             loading: true,
             numDelegatiors: 0,
-            delegations: ''
+            delegationsCount: 0
         }
     }
 
@@ -25,22 +25,7 @@ export default class ValidatorDelegations extends Component{
             }
 
             if (result){
-                // console.log(result);
-                // Delegations.remove({});
-                let Delegations = new Mongo.Collection(null);
-                result.forEach((delegation,i) => {
-                    Delegations.insert(delegation);
-                })
-                let delegations = Delegations.find({},{sort:{shares:-1}}).fetch();
-                this.setState({
-                    loading: false,
-                    numDelegatiors:delegations.length,
-                    delegations: delegations.map((d, i) => {
-                        return <Row key={i} className="delegation-info">
-                            <Col md={8} className="text-nowrap overflow-auto"><Account address={d.delegation.delegator_address} /></Col>
-                            <Col md={4}>{new Coin((d.balance.amount), d.balance.denom).stakeString()}</Col>
-                        </Row>
-                    })
+                this.setState({delegationsCount: result, loading: false
                 })
             }
         })
@@ -52,14 +37,13 @@ export default class ValidatorDelegations extends Component{
         }
         else{
             return <Card>
-                <CardHeader>{(this.state.numDelegatiors > 0)?this.state.numDelegatiors:'No'} <T>common.delegators</T> {(this.state.numDelegatiors > 0)?<small className="text-secondary">({new Coin(this.props.tokens/this.state.numDelegatiors).toString(2)}s / delegator)</small>:''}</CardHeader>
+                <CardHeader><T>validators.delegations</T></CardHeader>
                 <CardBody className="list">
                     <Container fluid>
                         <Row className="header text-nowrap d-none d-lg-flex">
-                            <Col md={8}><i className="fas fa-at"></i> <span><T>common.addresses</T></span></Col>
-                            <Col md={4}><i className="fas fa-piggy-bank"></i> <span><T>common.amounts</T></span></Col>
+                            <Col md={8}><span className="ml-n3"><T>common.totalNumOfDelegations</T>:</span></Col>
+                            <Col md={4}><span className="font-weight-normal float-right">{this.state.delegationsCount}</span></Col>
                         </Row>
-                        {this.state.delegations}
                     </Container>
                 </CardBody>
             </Card>