+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [UNRELEASED]
+* [#484] Replaced delegation list displayed under validator page with total number of delegations. 
+
 ## v0.41.x-12
 * [#387] Added Bluetooth Ledger support
 * Fix transaction simulation bug
@@ -19,7 +22,6 @@
 * Fixed an issue on displaying individual transaction.
 * Updated Ledger app name checking so that it will follows the value defined in settings.
 * [#213] Updated `@ledgerhq/hw-transport-webusb` pkg to v5.49.0 to fix Ledger errors on Windows10
-
 ## v0.41.x-7 (Stargate compatible)
 
 * [#472] Fix missing transactions